requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pymupdf
python-docx
scikit-learn
spacy
nltk
passlib[bcrypt]
python-jose[cryptography]
